Add Firebase integration and enhance event details in HTML
