+++ index.js
@@ -1,18 +1,17 @@
 // Import stylesheets
 import './style.css';
-// Firebase App (the core Firebase SDK) is always required
-import { initializeApp } from 'firebase/app';
 
-// Add the Firebase products and methods that you want to use
+// Import Firebase services from firebase.js
+import { auth, db } from './firebase';
+
+// Firebase authentication and Firestore services
 import {
-  getAuth,
   EmailAuthProvider,
   signOut,
   onAuthStateChanged,
 } from 'firebase/auth';
 
 import {
-  getFirestore,
   addDoc,
   collection,
   query,
@@ -28,7 +27,6 @@ import * as firebaseui from 'firebaseui';
 // Document elements
 const startRsvpButton = document.getElementById('startRsvp');
 const guestbookContainer = document.getElementById('guestbook-container');
-
 const form = document.getElementById('leave-message');
 const input = document.getElementById('message');
 const guestbook = document.getElementById('guestbook');
@@ -39,53 +37,39 @@ const rsvpNo = document.getElementById('rsvp-no');
 let rsvpListener = null;
 let guestbookListener = null;
 
-let db, auth;
-
 // Listen to guestbook updates
 function subscribeGuestbook() {
   const q = query(collection(db, 'guestbook'), orderBy('timestamp', 'desc'));
   guestbookListener = onSnapshot(q, (snaps) => {
-    // Reset page
     guestbook.innerHTML = '';
-    // Loop through documents in database
     snaps.forEach((doc) => {
-      // Create an HTML entry for each document and add it to the chat
       const entry = document.createElement('p');
-      entry.textContent = doc.data().name + ': ' + doc.data().text;
+      entry.textContent = `${doc.data().name}: ${doc.data().text}`;
       guestbook.appendChild(entry);
     });
   });
 }
 
-// Unsubscribe from guestbook updates
 function unsubscribeGuestbook() {
-  if (guestbookListener != null) {
+  if (guestbookListener) {
     guestbookListener();
     guestbookListener = null;
   }
 }
 
-// Listen for attendee list
 function subscribeCurrentRSVP(user) {
   const ref = doc(db, 'attendees', user.uid);
   rsvpListener = onSnapshot(ref, (doc) => {
     if (doc && doc.data()) {
       const attendingResponse = doc.data().attending;
-
-      // Update css classes for buttons
-      if (attendingResponse) {
-        rsvpYes.className = 'clicked';
-        rsvpNo.className = '';
-      } else {
-        rsvpYes.className = '';
-        rsvpNo.className = 'clicked';
-      }
+      rsvpYes.className = attendingResponse ? 'clicked' : '';
+      rsvpNo.className = attendingResponse ? '' : 'clicked';
     }
   });
 }
 
 function unsubscribeCurrentRSVP() {
-  if (rsvpListener != null) {
+  if (rsvpListener) {
     rsvpListener();
     rsvpListener = null;
   }
@@ -94,152 +78,73 @@ function unsubscribeCurrentRSVP() {
 }
 
 async function main() {
-  // Add Firebase project configuration object here
-  const firebaseConfig = {
-    apiKey: 'AIzaSyBN9oBxZC77p7pGRa1NPaHkwPB3nITRrzw',
-    authDomain: 'fir-web-codelab-22aeb.firebaseapp.com',
-    projectId: 'fir-web-codelab-22aeb',
-    storageBucket: 'fir-web-codelab-22aeb.firebasestorage.app',
-    messagingSenderId: '190908663044',
-    appId: '1:190908663044:web:a94ddce5f3484603827c51',
-  };
-
-  // Make sure Firebase is initilized
-  try {
-    if (firebaseConfig && firebaseConfig.apiKey) {
-      initializeApp(firebaseConfig);
-    }
-    db = getFirestore();
-    auth = getAuth();
-  } catch (e) {
-    console.log('error:', e);
-    document.getElementById('app').innerHTML =
-      '<h1>Welcome to the Codelab! Add your Firebase config object to <pre>/index.js</pre> and refresh to get started</h1>';
-    throw new Error(
-      'Welcome to the Codelab! Add your Firebase config object from the Firebase Console to `/index.js` and refresh to get started'
-    );
-  }
-
-  // FirebaseUI config
+  // FirebaseUI configuration
   const uiConfig = {
     credentialHelper: firebaseui.auth.CredentialHelper.NONE,
-    signInOptions: [
-      // Email / Password Provider.
-      EmailAuthProvider.PROVIDER_ID,
-    ],
+    signInOptions: [EmailAuthProvider.PROVIDER_ID],
     callbacks: {
-      signInSuccessWithAuthResult: function (authResult, redirectUrl) {
-        // Handle sign-in.
-        // Return false to avoid redirect.
-        return false;
-      },
+      signInSuccessWithAuthResult: () => false,
     },
   };
 
-  const ui = new firebaseui.auth.AuthUI(getAuth());
+  const ui = new firebaseui.auth.AuthUI(auth);
 
   // Listen to RSVP button clicks
   startRsvpButton.addEventListener('click', () => {
     if (auth.currentUser) {
-      // User is signed in; allows user to sign out
       signOut(auth);
     } else {
-      // No user is signed in; allows user to sign in
       ui.start('#firebaseui-auth-container', uiConfig);
     }
   });
 
-  // Listen to the current Auth state
-  // Listen to the current Auth state
+  // Listen to authentication state changes
   onAuthStateChanged(auth, (user) => {
     if (user) {
       startRsvpButton.textContent = 'LOGOUT';
-      // Show guestbook to logged-in users
       guestbookContainer.style.display = 'block';
-
-      // Subscribe to the guestbook collection
       subscribeGuestbook();
-      // Subscribe to the user's RSVP
       subscribeCurrentRSVP(user);
     } else {
       startRsvpButton.textContent = 'RSVP';
-      // Hide guestbook for non-logged-in users
       guestbookContainer.style.display = 'none';
-      // Unsubscribe from the guestbook collection
       unsubscribeGuestbook();
-      // Unsubscribe from the guestbook collection
       unsubscribeCurrentRSVP();
     }
   });
 
-  // Listen to the form submission
+  // Listen to form submission
   form.addEventListener('submit', async (e) => {
-    // Prevent the default form redirect
     e.preventDefault();
-    // Write a new message to the database collection "guestbook"
-    addDoc(collection(db, 'guestbook'), {
+    await addDoc(collection(db, 'guestbook'), {
       text: input.value,
       timestamp: Date.now(),
       name: auth.currentUser.displayName,
       userId: auth.currentUser.uid,
     });
-    // clear message input field
     input.value = '';
-    // Return false to avoid redirect
-    return false;
-  });
-
-  // Create query for messages
-  const q = query(collection(db, 'guestbook'), orderBy('timestamp', 'desc'));
-  onSnapshot(q, (snaps) => {
-    // Reset page
-    guestbook.innerHTML = '';
-    // Loop through documents in database
-    snaps.forEach((doc) => {
-      // Create an HTML entry for each document and add it to the chat
-      const entry = document.createElement('p');
-      entry.textContent = doc.data().name + ': ' + doc.data().text;
-      guestbook.appendChild(entry);
-    });
   });
 
   // Listen to RSVP responses
   rsvpYes.onclick = async () => {
-    // Get a reference to the user's document in the attendees collection
     const userRef = doc(db, 'attendees', auth.currentUser.uid);
-
-    // If they RSVP'd yes, save a document with attendi()ng: true
-    try {
-      await setDoc(userRef, {
-        attending: true,
-      });
-    } catch (e) {
-      console.error(e);
-    }
+    await setDoc(userRef, { attending: true });
   };
 
   rsvpNo.onclick = async () => {
-    // Get a reference to the user's document in the attendees collection
     const userRef = doc(db, 'attendees', auth.currentUser.uid);
-
-    // If they RSVP'd yes, save a document with attending: true
-    try {
-      await setDoc(userRef, {
-        attending: false,
-      });
-    } catch (e) {
-      console.error(e);
-    }
+    await setDoc(userRef, { attending: false });
   };
 
-  // Listen for attendee list
+  // Listen to attendee count updates
   const attendingQuery = query(
     collection(db, 'attendees'),
     where('attending', '==', true)
   );
-  const unsubscribe = onSnapshot(attendingQuery, (snap) => {
+  onSnapshot(attendingQuery, (snap) => {
     const newAttendeeCount = snap.docs.length;
-    numberAttending.innerHTML = newAttendeeCount + ' people going';
+    numberAttending.innerHTML = `${newAttendeeCount} people going`;
   });
 }
+
 main();
